third_party/blink/web_tests/editing/selection/select-from-textfield-outwards-expected.txt

start mac
PASS selection() is "0,0"
move at start of word
PASS selection() is "6,6"
move to end of word
PASS selection() is "6,11"
move to left
PASS selection() is "0,6"
move to right
PASS selection() is "6,17"
way up
PASS selection() is "0,6"
move to up
PASS selection() is "0,6"
move to down
PASS selection() is "6,17"
way down
PASS selection() is "6,17"
finish mac
PASS selection() is "6,17"
start unix
PASS selection() is "0,0"
move at start of word
PASS selection() is "6,6"
move to end of word
PASS selection() is "6,11"
move to left
PASS selection() is "0,6"
move to right
PASS selection() is "6,17"
way up
PASS selection() is "0,6"
move to up
PASS selection() is "0,6"
move to down
PASS selection() is "6,17"
way down
PASS selection() is "6,17"
finish unix
PASS selection() is "6,17"
start win
PASS selection() is "0,0"
move at start of word
PASS selection() is "6,6"
move to end of word
PASS selection() is "6,11"
move to left
PASS selection() is "0,6"
move to right
PASS selection() is "6,17"
way up
PASS selection() is "6,8"
move to up
PASS selection() is "6,8"
move to down
PASS selection() is "6,8"
way down
PASS selection() is "6,8"
finish win
PASS selection() is "6,8"
PASS successfullyParsed is true

TEST COMPLETE
